[Titin] add function conver 1IDR to 0.05INR, 100IDR to 0.5INR

+++ app/src/main/java/co/creativev/bootcamp/got/CurrencyConverter.java
@@ -2,12 +2,13 @@ package co.creativev.bootcamp.got;
 
 
 public class CurrencyConverter {
-    public int currencyConverter(int amount, String s) {
+    public double currencyConverter(double amount, String s) {
         if (s.equals("INR")){
             return amount*200;
         } else {
             return amount/200;
         }
     }
+
 
 }
